FEAT: 계모임 계좌에 잔액 충전하기 기능 구현

+++ src/main/java/com/team/gyemoim/dto/MyPageDTO.java
@@ -10,7 +10,7 @@ public class MyPageDTO {
   private String phone; //전화번호
   private String bankName; //은행명
   private String bankAccountNumber; //계좌번호
-  private double creditRating; //신용등급
+  private Integer creditRating; //신용등급
   private String enrollDate; //회원가입일
 
 }
